requirements.txt
boto3==1.34.21
python-dotenv==1.0.0
requests==2.32.5
pandas==2.3.3
openpyxl==3.1.5
xlrd==2.0.2
flask==3.1.0
